------------------------------------------------------------------------ r2772 | abbjo03 | 2010-04-08 16:30:10 +0100 (Thu, 08 Apr 2010) | 8 lines
Cross-integrate change 503530 from ingres2006r3.
The DAS symbol table is not distributed in the VMS debug
subdirectory.

    Add iigcd.stb to das.ccpp for distribution
    on VMS.

------------------------------------------------------------------------
r2773 | thogo01 | 2010-04-08 23:00:14 +0100 (Thu, 08 Apr 2010) | 30 lines

HOQA BE API tests produced DIFFs when run on a big-endian platform.
No DIFFs occur on little-endian platforms.  The DIFFs occur in COPY
values and are caused by copying GCA 4-byte integers into API 2-byte
values.  On little-endian plagforms, the low-order bytes of these
integers are the first bytes and therefore get copied into the
correct place.  The high order bytes are incorrectly copied and
overwrite the following values, but no problem occurs because the
values are copied in order and the overwrite occurs prior to the
following value being copied.  On big-endian platforms, only the
high-order bytes are copied into the destination values resulting
in incorrect values.  Intermediate 4-byte integer storage is used
for the copy, followed by a casted assignment to assign the correct
low-order bytes.

Batch tests showed a problem in API in detecting session transaction
state changes due to changes in message buffering.  Statement level
processing reads the GCA_RESPONSE message first and consumes the buffer.
Subsequent transaction level processing is not able to read the response
and verify the session transaction state.  API message processing has
the ability to read response messages without consuming the buffer,
which is done during event evaluation, as opposed to consuming the
message as is done during event processing.  Changed statement level
processing to not consume the response message so that it will be
accessible at the transaction level.

Fixed a typo which used the wrong protocol level for flagging database
procedure parameters as OUT params.

------------------------------------------------------------------------