------------------------------------------------------------------------ r2789 | stial01 | 2010-04-14 20:15:11 +0100 (Wed, 14 Apr 2010) | 17 lines
SIR 121123 Long Ids

Previous change for Long Ids (503565) caused problems selecting a
blob column in terminal monitor.

Changed terminal monitor to init new col_maxname before each query and before
each column is processed during a select.

Also made a minor fix to TRdisplay in the server. The DB_ATTS no longer holds
the attribute name, instead it contains a pointer to the attribute name.
As far as I can tell there is no valid format specifier (inside a repeating
group) for address of pointer. (It expects it to be the data).
Changed TRdisplay to just print the attid instead. Alternate fix is to loop
through and TRdisplay each attribute)

------------------------------------------------------------------------