------------------------------------------------------------------------ r2790 | bonro01 | 2010-04-14 21:45:11 +0100 (Wed, 14 Apr 2010) | 6 lines
A hang can occur if MOattach calls MO_alloc which needs to call MEget_pages to
allocate more pages. This results in a call to MOclassdef which also locks the
MO_mutex. To eliminate this problem the MO_mutex is released before calling
MO_alloc and re-acquired afterwards.

------------------------------------------------------------------------
r2791 | hweho01 | 2010-04-14 23:30:10 +0100 (Wed, 14 Apr 2010) | 13 lines

In LGposition(), the read log buffer pointed by context->lgc_buffer is set to a
memory that has to be aligned by a specific value which is returned from
DIget_direct_align(). However, when the buffer (LCTX) is allocated in
dm0l_allocate(), the value of align is not set by using the same function.  The
inconsistency causes memory overlay in some case, such as the dmfacp startup
problem on AIX platform.

To fix the error, the value of svcb_directio_align will be setup by calling
DIget_direct_align() in dmf_setup_directio() which will be added into the
dma_prepare() function. So the alignment size is set properly during the
initialization time for dmfacp.

------------------------------------------------------------------------