------------------------------------------------------------------------ r2791 | hweho01 | 2010-04-14 23:30:10 +0100 (Wed, 14 Apr 2010) | 13 lines
In LGposition(), the read log buffer pointed by context->lgc_buffer is set to a
memory that has to be aligned by a specific value which is returned from
DIget_direct_align(). However, when the buffer (LCTX) is allocated in
dm0l_allocate(), the value of align is not set by using the same function.  The
inconsistency causes memory overlay in some case, such as the dmfacp startup
problem on AIX platform.

To fix the error, the value of svcb_directio_align will be setup by calling
DIget_direct_align() in dmf_setup_directio() which will be added into the
dma_prepare() function. So the alignment size is set properly during the
initialization time for dmfacp.

------------------------------------------------------------------------
r2792 | loera01 | 2010-04-15 17:45:15 +0100 (Thu, 15 Apr 2010) | 20 lines

ODBC applications using the Windows driver manager are redirected to use the
Unicode version of ODBC functions that use string arguments, regardless of
whether the original code specifies the Unicode version.  SQLExecDirect() and
SQLPrepare() are therefore executed as SQLExecDirectW() and SQLPrepareW(),
respectively.

The Ingres ODBC driver reverts the Unicode strings back to multi-byte, and
uses the API function IIapi_convertData() to perform the conversion.  Since
the source and destination strings are described as an IIAPI_DATAVALUE, the
string lengths are u_i2.  The max value of a u_i2 is only 65535, or
approximately 64K.  For UCS2 (two-byte) Unicode strings, this means that
queries of only 32K can be converted at one time.

I revised the ExecDirectOrPrepareW() function, which converts the query
strings to multi-byte, to look for strings greater than 64 bytes in length.
If this threshhold is exceeded, ExecDirectOrPrepareW() converts the query
string in 64-byte segments before calling SQLExecDirect_InternalCall() or
SQLPrepare_InternalCall().

------------------------------------------------------------------------