------------------------------------------------------------------------ r2792 | loera01 | 2010-04-15 17:45:15 +0100 (Thu, 15 Apr 2010) | 20 lines
ODBC applications using the Windows driver manager are redirected to use the
Unicode version of ODBC functions that use string arguments, regardless of
whether the original code specifies the Unicode version.  SQLExecDirect() and
SQLPrepare() are therefore executed as SQLExecDirectW() and SQLPrepareW(),
respectively.

The Ingres ODBC driver reverts the Unicode strings back to multi-byte, and
uses the API function IIapi_convertData() to perform the conversion.  Since
the source and destination strings are described as an IIAPI_DATAVALUE, the
string lengths are u_i2.  The max value of a u_i2 is only 65535, or
approximately 64K.  For UCS2 (two-byte) Unicode strings, this means that
queries of only 32K can be converted at one time.

I revised the ExecDirectOrPrepareW() function, which converts the query
strings to multi-byte, to look for strings greater than 64 bytes in length.
If this threshhold is exceeded, ExecDirectOrPrepareW() converts the query
string in 64-byte segments before calling SQLExecDirect_InternalCall() or
SQLPrepare_InternalCall().

------------------------------------------------------------------------
r2793 | bonro01 | 2010-04-15 22:25:56 +0100 (Thu, 15 Apr 2010) | 1 line

Fix Usage error when calling set_env.sh
------------------------------------------------------------------------