------------------------------------------------------------------------ r2795 | smeke01 | 2010-04-16 09:30:36 +0100 (Fri, 16 Apr 2010) | 6 lines
Cross integration of change 503554

Code for MAXPAGE had typo adh_cost_estimate instead of ahd_page_estimate. Corrected this.

------------------------------------------------------------------------
r2796 | hanal04 | 2010-04-16 11:45:21 +0100 (Fri, 16 Apr 2010) | 12 lines

Added a -m (MustLog) flag to createdb so that a database can
be marked as MustLog (set nologging is blocked).
This change is in addition to the existing ability to make
a database when accessed by a given DBMS class (if it is in
the DBMS's MustLog list in CBF).
Added a "Always logged" status line to infodb so that the
user can clearly see whether the database is a MustLog database.
Added -enable_mustlog and -disable_mustlog options to alterdb
so that the user can update the MustLog status after  database
creation.

------------------------------------------------------------------------