------------------------------------------------------------------------ r2796 | hanal04 | 2010-04-16 11:45:21 +0100 (Fri, 16 Apr 2010) | 12 lines
Added a -m (MustLog) flag to createdb so that a database can
be marked as MustLog (set nologging is blocked).
This change is in addition to the existing ability to make
a database when accessed by a given DBMS class (if it is in
the DBMS's MustLog list in CBF).
Added a "Always logged" status line to infodb so that the
user can clearly see whether the database is a MustLog database.
Added -enable_mustlog and -disable_mustlog options to alterdb
so that the user can update the MustLog status after  database
creation.

------------------------------------------------------------------------
r2797 | hanal04 | 2010-04-16 16:15:13 +0100 (Fri, 16 Apr 2010) | 5 lines

Added dbmsmlst.hlp to the distribution. This was missed
in the original code submission during the code sprint
last year.

------------------------------------------------------------------------