------------------------------------------------------------------------ r2803 | bonro01 | 2010-04-16 23:15:19 +0100 (Fri, 16 Apr 2010) | 6 lines
After the rename of the vectorwise builds, the RPM savesets have a new prefix
of ingresvw.  Installations with these prefixes are not detected by standard
Ingres installations so they are not queried to determine which Installation
IDs are being used in these and prevent them from being re-used.

------------------------------------------------------------------------
r2804 | drivi01 | 2010-04-17 00:30:12 +0100 (Sat, 17 Apr 2010) | 3 lines

Insert NL at the end of the file to fix rcompare issues in reviewboard.

------------------------------------------------------------------------