------------------------------------------------------------------------ r2812 | jenjo02 | 2010-04-19 21:30:26 +0100 (Mon, 19 Apr 2010) | 6 lines
An access violation will occur in dm1p:unfix_header() if a FMAP has
not been fixed or locked.

Check that fmap_lock is not NULL before dereferencing it.

------------------------------------------------------------------------
r2813 | drivi01 | 2010-04-19 22:15:10 +0100 (Mon, 19 Apr 2010) | 13 lines

The Java Frequent Flyer demo can not be included with the distrubtion
due to the size of the application and its source.

The demo has been posted on community wiki and is available
through wiki examples page.

This change updates the document to point to the examples
page to help customers get started with application writing and
additional provides links to other pages within wiki such as
workshops, articles, presentations, FAQ and etc.

------------------------------------------------------------------------