------------------------------------------------------------------------ r2813 | drivi01 | 2010-04-19 22:15:10 +0100 (Mon, 19 Apr 2010) | 13 lines
The Java Frequent Flyer demo can not be included with the distrubtion
due to the size of the application and its source.

The demo has been posted on community wiki and is available
through wiki examples page.

This change updates the document to point to the examples
page to help customers get started with application writing and
additional provides links to other pages within wiki such as
workshops, articles, presentations, FAQ and etc.

------------------------------------------------------------------------
r2814 | horda03 | 2010-04-20 15:30:16 +0100 (Tue, 20 Apr 2010) | 34 lines

Cross integration of change 503760

n a Clustered Ingres installation, if the TX logs files contain
transactions that need recovery then the first DMFRCP server
started may take and never release a DB_OPEN and JOURNAL lock
on a DATABASE.

This occus when the TX's being recovered are in the local TX log file.

In a Clustered Ingres environment the recovery operation is different
from the non-clustered case. When Ingres isn't clustered, the RCP
server opens the (local) TX log file, takes locks on the DB's that
need recovery then defer the completion of the recovery to the ACP server.
The ACP server writes the TX log records to the journal (if appropriate)
then issues a LG_A_DBRECOVER - which causes the RCP server to release
the locks.

In a Clustered Installation, the RCP server has to do the ACP job because
it may be processing TX records from the TX log file of other nodess.
In this case, the recovery of the local TX's progress in the same way as
the non-clustered versin, with the exception that the RCP server does the
Archiving of the records - but the Archiving code in the RCP simply CLOSEd
the DB rather than signalling DBRECOVER; hence the DB_OPEN and JOURNAL locks
remain. For TX in other log files, the logging system is opened with
"LFB_USE_DIIO" flag set (as the nodename is specified in the call to LGopen). Thus
when the DB being recovered is added to the logging system (LGadd), the
LGD_OPENDB event isn't signalled.

The fact that the JOURNAL lock may persist for the lifetime of the Installation
means that the "locally" recovered DB's cannot be ckpdb on any other node; they
hang trying to get the JOURNAL lock.

------------------------------------------------------------------------