------------------------------------------------------------------------ r2815 | smeke01 | 2010-04-20 15:30:35 +0100 (Tue, 20 Apr 2010) | 10 lines
Cross integration of change 503621

Amended adc_compare so that it handles comparisons between nullable and
non-nullable values of the same type (instead of rejecting them with
error E_AD3001_DTS_NOT_SAME). Improved the error handling of the call
to adc_compare in pslsgram.yi so that a psf internal error and the text
of the adf error is logged.

------------------------------------------------------------------------
r2816 | maspa05 | 2010-04-20 15:30:43 +0100 (Tue, 20 Apr 2010) | 5 lines

Fix for 122642 introduced a mutex to protect UUID creation. ID_UUID_SEM_INIT
was defined to CS_synch_init on unix. Changed it to CS_cp_synch_init instead
so that it work across processes.

------------------------------------------------------------------------