------------------------------------------------------------------------ r2816 | maspa05 | 2010-04-20 15:30:43 +0100 (Tue, 20 Apr 2010) | 5 lines
Fix for 122642 introduced a mutex to protect UUID creation. ID_UUID_SEM_INIT
was defined to CS_synch_init on unix. Changed it to CS_cp_synch_init instead
so that it work across processes.

------------------------------------------------------------------------
r2817 | maspa05 | 2010-04-20 15:31:01 +0100 (Tue, 20 Apr 2010) | 56 lines

Cross integration of change 503678

This IP was originally 1121 but then the change was done in 9.3.
(see http://reviewboard.ingres.prv/r/1121/ for history)

This change makes sure that physical locks are used consistently on iisequence.
This was happening when updates were made but not during rollback.

dm2r_rcb_allocate sets rcb_k_duration to RCB_K_PHYSICAL if the tcb relates to
an S_CONCUR table (iirelation, iirel_idx, iiattribute, iiindex, iidevices)
or the reltid is DM_B_SEQ_TAB_ID (iisequence). dm0p_trans_page_lock checks
rcb_k_duration and if it's RCB_K_PHYSICAL sets the locking flag to LK_PHYSICAL.
This means we take physical locks during the row insert/update/delete itself.

However in order for physical locks to be taken during undo the log record
needs to have DM0L_PHYS_LOCK flag set. This gets set if tcb_sysrel is TRUE (see
put, replace, delete in dm1r) however tcb_sysrel was getting set only for
S_CONCUR catalogs.

Added a new relstat2 flag TCB2_PHYSLOCK_CONCUR to indicate catalogs that require
physical locks. Added this flag during the create of iisequence as well as the
existing S_CONCUR tables.

Changed the test in dm2t_alloc_tcb to set tcb_sysrel if it's
TCB2_PHYSLOCK_CONCUR.

During testing of this fix it's come to light that if we allow physical locks
on any structure but HASH it can cause problems. If for e.g. the row moves or
we have a btree split etc then session A may insert or update a row and then
release the lock meanwhile session B may do an update that cause the row to
move making session A's log record invalid. If session A then needs to rollback
we have a problem.

So changed dm2u_modify and where we disallow modify of S_CONCUR catalogs to
anything other than HASH added TCB2_PHYSLOCK_CONCUR catalogs to this
restriction.

Also amended upgradedb and verifydb to provide a method to check/fix that the
new flag is present where it needs to be and to set it. I've generalized this
to a defined list of "mandatory" flags for relstat and relstat2 for a list
of catalogs. Currently this list contains only iisequence and
TCB2_PHYSLOCK_CONCUR. Verifydb in runinteractive mode on the dbms_catalogs
will set the mandatory flags. It will also re-modify any TCB2_PHYSLOCK_CONCUR
tables to default structure (which will be HASH).

Similarly upgradedb will add the mandatory flags and re-modify any PHYSLOCK
tables.

In order to allow upgradedb and verifydb to call duc_modify_catalog on iirel_idx
I've added Duc_equivcats a list of catalog 'equivalents' for the purposes of
modify i.e. modifying iirelation effectively modifies iirel_idx. Amended
duc_modify_catalog to look for an equivalent and use that catalog's modify
statement instead if found.

------------------------------------------------------------------------