------------------------------------------------------------------------ r2835 | drivi01 | 2010-04-21 23:15:12 +0100 (Wed, 21 Apr 2010) | 9 lines
This change fixes link issues on Windows as a result of change #503771.
The change flips a compile flag on ducommon.qsh from REFERENCE_IN_DLL to
IMPORT_DLL_DATA to avoid symbol mismatch due to different type of
import/export.  ducommon.qsh symbols were being imported as
__declspec(dllimport) instead of extern.

The flag change fixes the linking problem.

------------------------------------------------------------------------
r2836 | loera01 | 2010-04-22 00:45:10 +0100 (Thu, 22 Apr 2010) | 42 lines

Change 503054, for bug 123318, replaced the scape of the GCN files with API
calls to the GCN for the ODBC configuration page.  As in change 503054, the
FileDoesExist(), GetNextVnode(), and FixComputerName() functions are
obsoleted.  These routines managed the read of the GCN login and vnode files.

The WM_INITDIALOG case in ConProcDriver() initializes the queue, which points
at a new VNODE_LIST structure, and the API.  The new BuildVnodeList()function
connects to the Name Server and calls the function BuildVnodeQueue() to fetch
vnode information and insert the information into the queue.

BuildVnodeList() differs slightly from its configuration counterpart in that
there is no need to initialize and terminate the API environment.  A valid API
handle already exists, which was created when AllocEnv() was called.  So,
BuildVnodeList() has an extra argument for the API environment handle.

After the vnode queue is created, the data must be qualified for the
VnodeSetsPromptUID() function, which is invoked after a vnode is selected.
If VnodeSetsPromptUID() determines that the vnode has no entry in the login
file, the "Prompt User ID and Password" box is automatically checked.  The
new function hasVnode() searches for at least one entry of a vnode name in
the login file.  A further check has to be made to the attributes file, since
a vnode may legitimatly have no login info if Kerberos is the authentication
mechanism.  So, hasVnode() is invoked for the attributes database as well.

The WM_DESTROY case empties the queue and frees any data allocated in
BuildVnodeList().

The previous approach had one advantage:  the GCN files could be read whether
the local GCN was running or not.  The API approach retrieves no vnodes if the
GCN is down.  BuildVnodeList() pops a MessagBox with a warning message in
this case, and gives the user the option of continuing or cancelling.  If the
user selects CANCEL, the data source configuration page is aborted via
PropSheet_PressButton().  If OK is selected, the user can enter a DSN
definition for local connections only, or modify information for data other
than vnodes.

One of these days, it would be nice to break these routines out of connect.c
and place them in a utility library that both the configuration and driver
routines could use.  There are some other functions duplicated in both modules
that could easily qualify.

------------------------------------------------------------------------