------------------------------------------------------------------------ r2851 | coomi01 | 2010-04-23 15:15:28 +0100 (Fri, 23 Apr 2010) | 6 lines
Cross integration of change 503813

Moved compile time #ifdef DOUBLEBYTE to 'if (CMGETDBL)' runtime test.

------------------------------------------------------------------------
r2852 | kiria01 | 2010-04-23 15:15:43 +0100 (Fri, 23 Apr 2010) | 28 lines

Correct application of aggregate join in sub-select

Background Information:

Reported by customer on II 9.3.1 (int.lnx/106)
Recreated here on head revs main

Documentation Impact:

None

When factoring a subselect into a derived table, aggregate restrictions
were being moved outside the aggregate to be a table restriction. This
has been reduced to only float predicates that are not completely satisfied
by the aggregate. Thus constant restrictions an aggregate local joins (as
in the test case) will be applied in the aggregation.

This change also cleans up the handling of the qualifier reducing code
needed to achieve it. This was not done this way originally due to the
perceived potential with multiple aggregates of adding multiple redundant
predicates but this will not affect the result and is probably eliminated
in OPF anyway.

Also tidied the DOT diagnostic code to avoid problems on VMS due to
accessing LOCATION private members outside of CL.

------------------------------------------------------------------------