------------------------------------------------------------------------ r2862 | drivi01 | 2010-04-23 23:15:10 +0100 (Fri, 23 Apr 2010) | 4 lines
Changes for bug #123385 remove dbmsmtpl.hlp from distribution.
This change removes this file from Windows distribution.

------------------------------------------------------------------------
r2863 | hanal04 | 2010-04-26 08:45:12 +0100 (Mon, 26 Apr 2010) | 15 lines

Cross integration of change 503922

GetTxnCvalue() returns the current value for a sequence.
If dms_seq->seq_cseq is NULL we search the TX's xcb_cseq
list for the sequence (DML_SEQ). If we find a match
we try to retrieve the next value. The problem is that
we were finding a stale reference to a dropped sequence
that had been recreated within the same TX.
This change clears down the cseq_seq reference in the
TX's xcb_cseq list when a sequence is dropped. selecting
seq.currval on the newly recreated sequence now fails
with E_QE00A2 as expected.

------------------------------------------------------------------------