------------------------------------------------------------------------ r2863 | hanal04 | 2010-04-26 08:45:12 +0100 (Mon, 26 Apr 2010) | 15 lines
Cross integration of change 503922

GetTxnCvalue() returns the current value for a sequence.
If dms_seq->seq_cseq is NULL we search the TX's xcb_cseq
list for the sequence (DML_SEQ). If we find a match
we try to retrieve the next value. The problem is that
we were finding a stale reference to a dropped sequence
that had been recreated within the same TX.
This change clears down the cseq_seq reference in the
TX's xcb_cseq list when a sequence is dropped. selecting
seq.currval on the newly recreated sequence now fails
with E_QE00A2 as expected.

------------------------------------------------------------------------
r2864 | smeke01 | 2010-04-26 16:00:15 +0100 (Mon, 26 Apr 2010) | 7 lines

Cross integration of change 503993

Added explicit derived_table production to enable a derived table
to appear on the right hand side of a join operator.

------------------------------------------------------------------------