------------------------------------------------------------------------ r2865 | drivi01 | 2010-04-26 16:30:27 +0100 (Mon, 26 Apr 2010) | 6 lines
Documentation changes are needed for Usability changes submitted
via SIR 123397.
This change also includes cbf parameter changes which bring
the tool in sync with documentation.

------------------------------------------------------------------------
r2866 | drivi01 | 2010-04-26 18:00:25 +0100 (Mon, 26 Apr 2010) | 10 lines

Currently main build is broken.  This is due to a recent change
to Jamrules that makes import/export IMPORT_DLL_DATA and
REFERENCE_IN_DLL flag rely on grist instead of girst and
filename, resulting in certain files being compiled with no
flag or the wrong flag.

As a result the import/export symbols are now mismatched
and we're seeing compilation issues.

------------------------------------------------------------------------