------------------------------------------------------------------------ r2866 | drivi01 | 2010-04-26 18:00:25 +0100 (Mon, 26 Apr 2010) | 10 lines
Currently main build is broken.  This is due to a recent change
to Jamrules that makes import/export IMPORT_DLL_DATA and
REFERENCE_IN_DLL flag rely on grist instead of girst and
filename, resulting in certain files being compiled with no
flag or the wrong flag.

As a result the import/export symbols are now mismatched
and we're seeing compilation issues.

------------------------------------------------------------------------
r2867 | schka01 | 2010-04-26 19:16:08 +0100 (Mon, 26 Apr 2010) | 13 lines

Implemented a BQCB (blob query context) to allow blobs
to know more about the surrounding context, so that
etabs will be opened in a compatible mode (table
locking, MVCC, etc).
Arrange for COPY to directly load etabs instead of
sending incoming blobs to holding temps first.
Require coupon context when looking at blobs;  fix
code (such as duplicate checking) that did not supply
coupon context to dmpe.
Short-circuit COPY in the sequencer so that in-
progress COPY messages are handler more directly.

------------------------------------------------------------------------