------------------------------------------------------------------------ r2872 | coomi01 | 2010-04-27 11:00:24 +0100 (Tue, 27 Apr 2010) | 12 lines
Cross integration of change 503049

Port change 501686 from main to 2006r2.

Direct X-integrate is not possible as port completes 65 catalogs
whereas change 501686 supports only 93 catalogs.

Will x-integrate this change up to main complementing existing
code when ready.

------------------------------------------------------------------------
r2873 | coomi01 | 2010-04-27 11:45:14 +0100 (Tue, 27 Apr 2010) | 12 lines

Expressions, which include function calls, are a bigger set
than qual_column_ref. In particular parameters are seperated by
commas which led to syntax errors in the original.

A problem here will be the semantics of using an expression,
which in general may not return a constant integer value.
However, the parser has the job of scanning syntax and placing
output in a string ready for the backend parser to compile. Hence,
though we are being more open than needs be at this stage any
errors will be captured and reported at a latter stage.

------------------------------------------------------------------------