------------------------------------------------------------------------ r2873 | coomi01 | 2010-04-27 11:45:14 +0100 (Tue, 27 Apr 2010) | 12 lines
Expressions, which include function calls, are a bigger set
than qual_column_ref. In particular parameters are seperated by
commas which led to syntax errors in the original.

A problem here will be the semantics of using an expression,
which in general may not return a constant integer value.
However, the parser has the job of scanning syntax and placing
output in a string ready for the backend parser to compile. Hence,
though we are being more open than needs be at this stage any
errors will be captured and reported at a latter stage.

------------------------------------------------------------------------
r2874 | kiria01 | 2010-04-27 12:30:36 +0100 (Tue, 27 Apr 2010) | 13 lines

Fix of b123632 unicode combining collation can hang dbms in cpu loop.

The combining entry information list for a combining lead character
was not completly iterated across - the last entry was skipped.
This code loop appears in 3 places in aduucmp.c and in the case of
adu_pat_MakeCElen() the skipping of the last block is critical if
it was the only block present as the source string pointer was not
advanced resulting in an infinite loop.

The loop end test has been corrected and the loop test further secured
against infinite looping by ensuring that zero match length
combinations are not treated as combinations.

------------------------------------------------------------------------