------------------------------------------------------------------------ r2874 | kiria01 | 2010-04-27 12:30:36 +0100 (Tue, 27 Apr 2010) | 13 lines
Fix of b123632 unicode combining collation can hang dbms in cpu loop.

The combining entry information list for a combining lead character
was not completly iterated across - the last entry was skipped.
This code loop appears in 3 places in aduucmp.c and in the case of
adu_pat_MakeCElen() the skipping of the last block is critical if
it was the only block present as the source string pointer was not
advanced resulting in an infinite loop.

The loop end test has been corrected and the loop test further secured
against infinite looping by ensuring that zero match length
combinations are not treated as combinations.

------------------------------------------------------------------------
r2875 | schka01 | 2010-04-27 13:00:11 +0100 (Tue, 27 Apr 2010) | 7 lines

Addition is not commutative when it's string concatenation;
Don't combine s1+s2 and s2+s1 when s1, s2 are strings.
Avoid common subexpression elimination entirely when it's a
blob expression, since DMF makes assumptions about blob holding
temp lifetime that CSE elimination can violate.

------------------------------------------------------------------------