------------------------------------------------------------------------ r2875 | schka01 | 2010-04-27 13:00:11 +0100 (Tue, 27 Apr 2010) | 7 lines
Addition is not commutative when it's string concatenation;
Don't combine s1+s2 and s2+s1 when s1, s2 are strings.
Avoid common subexpression elimination entirely when it's a
blob expression, since DMF makes assumptions about blob holding
temp lifetime that CSE elimination can violate.

------------------------------------------------------------------------
r2876 | coomi01 | 2010-04-27 13:30:09 +0100 (Tue, 27 Apr 2010) | 19 lines

Cross integration of change 502190

In xferdb writeview was semi state like, in assuming it would
receive correct and ordered results from a Union select.
When sent items out of order, it could SegV whould a previous
setup state have been skipped.

In the change, I have taken the stream setup out of the if body
and set it to work whenever there is a change of view. This includes
the first view, as no reltid may be -1.

This stops the SegV occuring, the point of this change.

Of course, if the dbms it not returning results as expected from
the query, there is no way xferdb will detect this. The resulting
copy.in file will be correct but incomplete.

------------------------------------------------------------------------