------------------------------------------------------------------------ r2881 | loera01 | 2010-04-27 20:30:13 +0100 (Tue, 27 Apr 2010) | 36 lines
This change adds iterative capability to SQLBrowseConnect().

I added a boolean argument, bcConnectCalled, to the tDBC structure to track
the first invocation of SQLBrowseConnect().  So, on the initial call,
SQLBrowseConnect() parses the input connection string, and returns a string
indicating all of the connection attributes required to connect to Ingres.
Optional attributes, such as UID/PWD or GROUP, are flagged with an asterisk.
Otherwise, the recognized portion of the input connection string is not
returned.

Each invocation of SQLBrowseConnect() looks for attributes not previously
specified, and stores them in the connection handle (tDBC) when found.  A
working connection string named bcOutStr is initialized according to the
specified length in SQLBrowseConnect().  Each pass of SQLBrowseConnect()
rebuilds bcOutStr according to discovered attributes.

Although SQLServer can apparently connect in more than two phases, the Ingres
ODBC driver requires only two at most.  The first phase requires the driver
name or DSN name to be specified in order for the the driver to be loaded.
This phase is not necessary in a CLI application, which simply loads the
driver from the appropriate location in II_SYSTEM.  For applications
specifying the DSN attribute, all other connection attributes are optional;
for applications specifying the Driver attribute and not DSN's, the
SERVERTYPE, SERVERNAME (vnode), and DATABASE attributes are required.
Otherwise, SQLBrowseConnect() returns SQL_NEED_DATA with the SQLSTATE 01S00.

Once SQLBrowseConnect() receives the minimum attributes, it invokes
SQLDriverConnect() with the connection string built from the previously-
specified attributes.  At that point, the output string returned to the
caller actually comes from SQLDriverConnect().

If SQLBrowseConnect() receives a value/attribute pair that it cannot
recognize, it returns SQL_NEED_DATA and SQLSTATE 01S00, even if the
connection requirements are otherwise met.

------------------------------------------------------------------------
r2882 | rajus01 | 2010-04-27 21:30:10 +0100 (Tue, 27 Apr 2010) | 5 lines

When testing the change 503953 on windows platform it was noticed that when the network listen ports are specified using the new port syntax (i.e of the form II7+ for example ), the iimonitor "show server" command failed to display the actual symbolic listen port for which the TCP/IP listen was completed. The behavior is same across GCB/GCC/GCD servers when using TCP/IP network protocol. The info was displayed correctly on Unix platforms due to my change for bug 120552 (Unix specific changes in the TCP/IP protocol driver).

Ported the Unix CL changes for bug 120552 to the TCP/IP driver(tcp_ip.c) to fix the port display issue on Windows platform.  The routine GCtcpip_port() now takes additional parameter to obtain the actual symbolic port from the GCtcpip_port() routine which maps the symbolic port with the numeric port. Assumed no changes to WINTCP protocol driver (win_tcp.c) is required as it was deprecated.

------------------------------------------------------------------------