------------------------------------------------------------------------ r2903 | kiria01 | 2010-04-29 07:12:57 +0100 (Thu, 29 Apr 2010) | 5 lines
 When checking the context of a SELECT token, nolonger assume that a SELECT
 found in an active FROM list implies a derived table as it could be a subselect
 in the target list, where clause or on clause.

------------------------------------------------------------------------
r2904 | coomi01 | 2010-04-29 08:15:32 +0100 (Thu, 29 Apr 2010) | 25 lines

In C we may define a constant expression and use it as an array dimension
within esqlc source code

#define I_DIM 41 + 1

   exec sql begin declare section;
    wchar_t fname[I_DIM];
   exec sql end declare section;

When emiting code for space allocation, esqlc will produce
	IIgetdomio((short *)0,1,26,(I_DIM*4)-4,fname);
Or equivalent.

The C compiler sees the size expression as

	41 + 1 * 4

Giving 46, not 168 as required.

The change brackets the dimstr to ensure, if it is an expression, it
gets evaluated before it is scaled.

------------------------------------------------------------------------