------------------------------------------------------------------------ r2917 | coomi01 | 2010-04-30 10:15:48 +0100 (Fri, 30 Apr 2010) | 10 lines
Change 503214 went futher than needed, and caused a problem
accessing sequences owned by dba by another user.

In this change, I have reverted the test to what it used to be
and then barred it being used whenever the mode is DSEQUENCE.

Now the user may find the dba sequence as long as they are not
trying to delete it.

------------------------------------------------------------------------
r2918 | loera01 | 2010-04-30 16:00:18 +0100 (Fri, 30 Apr 2010) | 9 lines

This change adds ODBC support for positional parameters in the DBMS for Ingres
10.

The internal function ScanProcParamsAndBuildDescriptor() calls
GetProcParamNames(), which performs a catalog query for the parameter names,
which are then applied to the IPD of the call to the database procedure.  I
made this call contingent on IIAPI_LEVEL_6.

------------------------------------------------------------------------