------------------------------------------------------------------------ r2918 | loera01 | 2010-04-30 16:00:18 +0100 (Fri, 30 Apr 2010) | 9 lines
This change adds ODBC support for positional parameters in the DBMS for Ingres
10.

The internal function ScanProcParamsAndBuildDescriptor() calls
GetProcParamNames(), which performs a catalog query for the parameter names,
which are then applied to the IPD of the call to the database procedure.  I
made this call contingent on IIAPI_LEVEL_6.

------------------------------------------------------------------------
r2919 | thoda04 | 2010-04-30 16:00:32 +0100 (Fri, 30 Apr 2010) | 4 lines

Add Ingres .NET Data Provider support for AddWithValue()
method in the IngresParameterCollection class.  Bug 123662.

------------------------------------------------------------------------